Just another addition for my GitHub repo
Tested this script today on Raspberry Pi 4 and it worked as expected. I still need to debug this script to find out why the smile output doesn't work the way it did, back in the day, but I'll save that for another time when I have a bit more time.